Allow god characters to use the "view position" feature
This teleports the view for players with no set character
(player.character = nil) rather than use ls-controller which would
lose the player inventory.

The drawback to this is that the player can still move the view when the
location marker is there, but this is not really an issue since the player is
a god anyway.